fix(storybook): resolve CI failures from story splits
- Remove unused imports/variables across 18 story files (TS6133/TS6192)
- Add missing within/expect imports in CreateProjectDialog, CommandPaletteList
- Remove dead SystemDialog wrappers from AiDialogsStates (moved to StatesSystem)
- Split ListItemAdvanced → ListItemMatrix to stay under 500-line guard
- Fix prettier formatting for all modified story files
- Remove unused EMPTY_DATA, NodeType, PopoverClose, useState refs
