+++ apps/desktop/renderer/src/components/primitives/TextVariants.stories.tsx
@@ -85,7 +85,6 @@ const colors: TextColor[] = [
   "info",
 ];
 
-
 // ============================================================================
 // 边界情况
 // ============================================================================
